add features in todo app
