+++ src/components/todo.css
@@ -1,0 +1,154 @@
+body {
+    font-family: Arial, sans-serif;
+    background-image: url(../images/background-img.jpg);
+    margin: 0;
+    padding: 0;
+    display: flex;
+    justify-content: center;
+    align-items: flex-start; 
+    min-height: 100vh;
+    overflow: hidden;
+    font-family: Arial, sans-serif;
+    background-size: cover; 
+    background-position: center; 
+    background-repeat: no-repeat; 
+    align-items: center;
+  }
+  
+  .main_container {
+    background-color: #fff;
+    border-radius: 8px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    width: 400px;
+    max-height: 90vh; 
+    display: flex;
+    flex-direction: column;
+    overflow: hidden; 
+  }
+  
+  #container {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+    flex-grow: 1;
+    overflow-y: auto; 
+    padding: 10px; 
+  }
+  
+  h1 {
+    text-align: center;
+    color: #333;
+    margin: 0;
+    padding: 10px 0; 
+    background-color: #f8f9fa; 
+    border-bottom: 1px solid #ddd; 
+  }
+  
+ 
+  #input_task {
+    padding: 10px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    font-size: 1rem;
+    width: calc(100% - 22px); 
+    margin-bottom: 10px; 
+  }
+  
+  #add_task_btn, #delete_complete_btn {
+    padding: 10px;
+    border: none;
+    border-radius: 4px;
+    background-color: #28a745;
+    color: white;
+    font-size: 1rem;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+    margin-bottom: 10px; 
+  }
+  
+  #add_task_btn:hover, #delete_complete_btn:hover {
+    background-color: #218838;
+  }
+  
+  #delete_complete_btn {
+    background-color: #dc3545;
+  }
+  
+  #delete_complete_btn:hover {
+    background-color: #c82333;
+  }
+  
+ 
+  .list-container {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    flex-grow: 1;
+    overflow-y: auto; 
+  }
+  
+  .list-container li {
+    background-color: #f8f9fa;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    padding: 10px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    word-break: break-word;
+  }
+  
+  .list-container li.completed {
+    background-color: #e9ecef;
+  }
+  
+  .list-container li.completed label {
+    text-decoration: line-through;
+    color: #dc3545;
+  }
+  
+  .list-container li label {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    flex-grow: 1;
+  }
+  
+  .list-container li input[type="checkbox"] {
+    cursor: pointer;
+  }
+  
+  #edit, #save {
+    padding: 5px 10px;
+    border: none;
+    border-radius: 4px;
+    background-color: #007bff;
+    color: white;
+    font-size: 0.8rem;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+  }
+  
+  #edit:hover, #save:hover {
+    background-color: #0056b3;
+  }
+  
+  #save {
+    background-color: #28a745;
+  }
+  
+  #save:hover {
+    background-color: #218838;
+  }
+  
+  .list-container li input[type="text"] {
+    padding: 5px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    font-size: 0.8rem;
+    flex-grow: 1;
+    margin-right: 10px;
+  }
